\name{girlgrowth}
\alias{girlgrowth}
\docType{data}
\title{
Growth data for girls
}
\description{
Heights of some adolescent girls, aged 7 to 12, in the southern part of Kolkata, India around the year 2008.
}
\usage{data(girlgrowth)}
\format{
  A data frame with 905 observations on the following 2 variables.
  \describe{
    \item{\code{Age}}{Age of girls (in years)}
    \item{\code{Height}}{Height of girls (in cm)}
  }
}

\source{
Dasgupta (2015), Physical Growth, Body Composition and Nutritional Status of Bengali School aged Children, Adolescents and Young adults of Calcutta, India: Effects of Socioeconomic Factors on Secular Trends, Report 158, Ney-van Hoogstraten Foundation, The Netherlands.
}
\examples{
data(girlgrowth)
head(girlgrowth)
}
\keyword{datasets}
