\name{waist}
\alias{waist}
\docType{data}
\title{
Waist circumference and adipose tissue data
}
\description{
Waist circumference and adipose tissue data (Daniel and Cross, 2013).
}
\usage{data(waist)}
\format{
  A data frame with 109 observations on the following 2 variables.
  \describe{
    \item{\code{Waist}}{Waist circumference (in centimeters)}
    \item{\code{AT}}{Area of lower abdominal adipose tissue (in squared centimeters)}
  }
}

\source{
Daniel, W.W. and Cross, C.L. (2013) Biostatistics: A Foundation for Analysis in the Health Sciences, tenth edition, Wiley, New York, Table 9.3.1.
}

\examples{
data(waist)
head(waist)
}
\keyword{datasets}
